requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
bidict==0.23.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
defusedxml==0.7.1
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
filelock==3.18.0
fsspec==2025.3.2
google-auth==2.39.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.9.0
oauthlib==3.2.2
openai==1.76.2
packaging==25.0
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
python3-openid==3.2.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
service-identity==24.2.0
setuptools==80.0.0
simple-websocket==1.1.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.0
sqlparse==0.5.3
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
whitenoise==6.9.0
wsproto==1.2.0
zope.interface==7.2
